Conflicts:
	mi/idk/comm_config.py
	mi/idk/unit_test.py

+++ mi/idk/dataset/unit_test.py
@@ -45,7 +45,7 @@ from mi.core.instrument.instrument_driver import DriverEvent
 from interface.objects import ResourceAgentConnectionLostErrorEvent
 
 from pyon.core.exception import Conflict
-from pyon.core.exception import ResourceError, BadRequest, Timeout, ServerError
+from pyon.core.exception import ResourceError, BadRequest, ServerError
 from pyon.agent.agent import ResourceAgentState
 from pyon.agent.agent import ResourceAgentEvent
 
@@ -257,7 +257,9 @@ class DataSetTestCase(MiIntTestCase):
         self.clear_sample_data()
         if isinstance(data_dir, list):
             for d_dir in data_dir:
-                os.rmdir(d_dir)
+                # confirm this path exists in case we configure two parsers to look at the same dir
+                if os.path.exists(d_dir):
+                    os.rmdir(d_dir)
         else:
             os.rmdir(data_dir)
 
@@ -285,7 +287,7 @@ class DataSetTestCase(MiIntTestCase):
             elif os.path.isdir(stored_data_dir):
                 remove_all_files(stored_data_dir)
 
-    def create_sample_data(self, filename, dest_filename=None, mode=0644, create=True):
+    def create_sample_data(self, filename, dest_filename=None, mode=0644, create=True, copy_metadata=True):
         """
         Search for a data file in the driver resource directory and if the file
         is not found there then search using the filename directly.  Then copy
@@ -297,6 +299,7 @@ class DataSetTestCase(MiIntTestCase):
         @param: dest_filename - name of the file when copied. default to filename
         @param: file mode
         @param: create an empty file in the destination if the source is not found
+        @param: copy_metadata - True to copy file metadata false to not copy metadata
         @return: path to file created
         """
         data_dir = self.create_data_dir()
@@ -322,13 +325,21 @@ class DataSetTestCase(MiIntTestCase):
             file = open(dest_path, 'w')
             file.close()
         else:
-            shutil.copy2(source_path, dest_path)
+            if copy_metadata:
+                # copy the file and its metadata
+                # this leaves the file modification time the same as the original file
+                shutil.copy2(source_path, dest_path)
+            else:
+                # copy the just the data
+                # this changes the file modification time to the time of the copy
+                shutil.copy(source_path, dest_path)
 
         os.chmod(dest_path, mode)
 
         return dest_path
 
-    def create_sample_data_set_dir(self, filename, dest_dir, dest_filename=None, mode=0644, create=True):
+    def create_sample_data_set_dir(self, filename, dest_dir, dest_filename=None,
+                                   mode=0644, create=True, copy_metadata=True):
         """
         Search for a data file in the driver resource directory and if the file
         is not found there then search using the filename directly.  Then copy
@@ -340,6 +351,7 @@ class DataSetTestCase(MiIntTestCase):
         @param: dest_filename - name of the file when copied. default to filename
         @param: file mode
         @param: create an empty file in the destination if the source is not found
+        @param: copy_metadata - True to copy file metadata false to not copy metadata
         @return: path to file created
         """
         if not os.path.exists(dest_dir):
@@ -370,7 +382,14 @@ class DataSetTestCase(MiIntTestCase):
             file = open(dest_path, 'w')
             file.close()
         else:
-            shutil.copy2(source_path, dest_path)
+            if copy_metadata:
+                # copy the file and its metadata
+                # this leaves the file modification time the same as the original file
+                shutil.copy2(source_path, dest_path)
+            else:
+                # copy the just the data
+                # this changes the file modification time to the time of the copy
+                shutil.copy(source_path, dest_path)
 
         os.chmod(dest_path, mode)
 
@@ -484,7 +503,7 @@ class DataSetIntegrationTestCase(DataSetTestCase):
                     log.debug("No exception detected yet, sleep for a bit")
                     gevent.sleep(1)
 
-        except Timeout:
+        except gevent.Timeout:
             log.error("Failed to detect exception %s", exception_class)
             self.fail("Exception detection failed.")
 
@@ -510,7 +529,7 @@ class DataSetIntegrationTestCase(DataSetTestCase):
                     log.debug("No event detected yet, sleep for a bit")
                     gevent.sleep(1)
 
-        except Timeout:
+        except gevent.Timeout:
             log.error("Failed to detect event %s", event_class_str)
             self.fail("Event detection failed.")
 
@@ -525,18 +544,20 @@ class DataSetIntegrationTestCase(DataSetTestCase):
         @param count, how many records to wait for
         @param timeout, how long to wait for the records.
         """
-        try:
-            particles = self.get_samples(particle_class, count, timeout)
-        except Timeout:
-            log.error("Failed to detect particle %s, expected %d particles, found %d", particle_class, count, found)
-            self.fail("particle detection failed. Expected %d, Found %d" % (count, found))
+        particles = self.get_samples(particle_class, count, timeout)
 
-        # Verify the data against the result data set definition
-        if result_set_file:
-            rs_file = self._get_source_data_file(result_set_file)
-            rs = ResultSet(rs_file)
+        if len(particles) == count:
+            # Verify the data against the result data set definition
+            if result_set_file:
+                rs_file = self._get_source_data_file(result_set_file)
+                rs = ResultSet(rs_file)
 
-            self.assertTrue(rs.verify(particles), msg="Failed data validation, check the logs.")
+                self.assertTrue(rs.verify(particles), msg="Failed data validation, check the logs.")
+        else:
+            log.error("%d particles were requested but only %d were found within the timeout of %d seconds",
+                      count, len(particles), timeout)
+            self.fail("%d particles were requested but only %d were found within the timeout of %d seconds" %
+                      (count, len(particles), timeout))
 
     def assert_file_ingested(self, filename, data_source_key=None):
         """
@@ -552,14 +573,17 @@ class DataSetIntegrationTestCase(DataSetTestCase):
         if not filename in last_state or not last_state[filename]['ingested']:
             self.fail("File %s was not ingested" % filename)
 
-    def assert_file_not_ingested(self, filename):
+    def assert_file_not_ingested(self, filename, data_source_key=None):
         """
         Assert that a particular file was not ingested (useable by Single Directory driver, not Single File driver),
         If the ingested flag is set in the driver state for this file, fail the test
         @ param filename name of the file to check that it was ingested using the ingested flag
         """
         log.debug("last state callback result %s", self.state_callback_result[-1])
-        last_state = self.state_callback_result[-1]
+        if data_source_key is None:
+            last_state = self.state_callback_result[-1]
+        else:
+            last_state = self.state_callback_result[-1][data_source_key]
         if filename in last_state and last_state[filename]['ingested']:
             self.fail("File %s was ingested when we expected it not to be" % filename)
 
@@ -587,7 +611,7 @@ class DataSetIntegrationTestCase(DataSetTestCase):
                     if particle_class is None or isinstance(data, particle_class):
                         found += 1
                         result.append(self.data_callback_result.pop(check_idx))
-                        log.debug("Found sample index %d, #%d", check_idx, found)
+                        log.trace("Found sample index %d, #%d", check_idx, found)
                     else:
                         # skip past a particle that doesn't match our particle class
                         check_idx += 1
@@ -603,7 +627,7 @@ class DataSetIntegrationTestCase(DataSetTestCase):
                 if not done and self.data_callback_result == []:
                     log.debug("No particle detected yet, sleep for a bit")
                     gevent.sleep(1)
-        except Timeout:
+        except gevent.Timeout:
             log.error("Failed to detect particle %s, expected %d particles, found %d", particle_class, count, found)
             result = []
         finally:
@@ -1040,7 +1064,7 @@ class DataSetAgentTestCase(DataSetTestCase):
                 if not done:
                     log.debug("state mismatch, waiting for state to transition.")
                     gevent.sleep(1)
-        except Timeout:
+        except gevent.Timeout:
             log.error("Failed to transition agent state to %s, current state: %s", target_agent_state, agent_state)
             self.fail("Failed to transition state.")
         finally:
@@ -1067,7 +1091,7 @@ class DataSetAgentTestCase(DataSetTestCase):
                 if not done:
                     log.debug("target event not detected, sleep a bit to let events happen")
                     gevent.sleep(1)
-        except Timeout:
+        except gevent.Timeout:
             log.error("Failed to find event in queue: %s", event_object_type)
             log.error("Current event queue: %s", self.event_subscribers._events_received)
             self.fail("%s event not detected")
@@ -1515,7 +1539,7 @@ class DataSetIngestionTestCase(DataSetAgentTestCase):
                 log.debug("In our event sleep loop.  just resting for a bit.")
                 gevent.sleep(sleeptime)
 
-        except Timeout:
+        except gevent.Timeout:
             log.info("Finished ingestion test as runtime has been exceeded")
 
         finally:
